Add job to CI that cuts release and uploads a variety of binaries
Add a release CI job that upon detection of new tag, builds binaries for
Darwin / Linux operating systems and `amd64` / `arm64` architectures and
cuts a GitHub release with them.

We only both with two operating systems and two architectures because in
practice, that's what 99.9% of people will be using at this point.

Most of the release job still runs even when not building a tag just so
it gets more regular exercise and is easier for us to iterate on when we
need to make changes to it.