Remove `.golangci.yml` in favor of `.golangci.yaml` (#40)
The repository currently has both a `.golangci.yml` and a
`.golangci.yaml` in place that are 95% the same. The only place they
were different is that one was disabling a gosec rule and the other
disabling a Revive rule. It seems that neither disable is necessary in
this project, so I just took them both out.