Add install instructions to the README
Change the README to consist mainly of a description of the project
along with install instructions with binaries and container images,
which will be more useful to most people looking at the project.

In sitting with the same convention as River, move development
instructions to `./docs/development.md` and also move the `README.md`
into `./docs` so that we can have separate documentation for the
top-level Go package.