Convert all queue endpoints + more tests + list response
Continue API endpoint conversion work by converting all queue endpoints
to the new system and add requisite tests for all of them.

Another thing I added here is a generic `listResponse` so that list
endpoints start returning top level objects instead of raw arrays. This
is more compatible with the API endpoint framework for one, but from an
API design standpoint it's better because it's more forward compatible.
In case you ever want to add a new field to a list response, you don't
have any flexibility on a raw array because there's nowhere to put
anything new, but with a top-level object, you can add new keys to it.