Match sqlc options of main River project (#70)
I'm finding it a little disorienting that the use of sqlc functions in
River UI are a little different than their use in the main River project
due to the options in play being a little different. e.g. DBTX as a
`New()` argument instead of one to the function itself, non-pointer
parameters and return values.

Here, match the sqlc options with the main River project, and update
call sites so everything works.