Prepare release 0.2.0 (#83)
It's been a while since we cut a River UI release and a couple new
features and changes have some in, so I figure it's not a bad idea to
send one one. Here, prepare version 0.2.0.

Also, add some release instructions. It'd be nice if these were more
automated, but it's better to have manual release instructions instead
of nothing for now.