Interpret Postgres connect errors to user-facing (#94)
Reveal a Postgres connect error as user-facing. This will be helpful for
operators in debugging a River UI installation, especially when first
setting it up.

We use the full error string from the Pgx connect error -- it contains
useful context and Pgx is smart enough to already sanitize it such that
it won't contain a password or other secrets.

Fixes #93.