Introduce middleware concept to API infrastructure (#95)
Here, put in a basic middleware concept to the API infrastructure. The
middleware interface is the same as the one found in packages like Mux:

    type middlewareInterface interface {
        Middleware(next http.Handler) http.Handler
    }

But another thing we do is bring in a construct called `MiddlewareStack`
that facilitates building a middleware stack in a more human-intuitive
way.

Calling using middleware functions only, middleware ends up being
combined like so:

    handler := mux
    handler = middleware1.Wrapper(handler)
    handler = middleware2.Wrapper(handler)

The trouble with this approach is that the middleware ends up getting
ordered unintuitively. It seems like `middleware1` should be what gets
run first because it's defined at the top, but it's actually the
middleware that runs _last_ because it's wrapped most deeply. You can
keep that mental model together when there's just a couple middleware,
but when you start building large stacks of middleware, it gets more
difficult that you need to think back-to-front.

So here we provide `MiddlewareStack` which works similarly, except when
mounting middlewares at the end, it inverts the order of the input
middleware, thereby allowing middleware to be added in an order that's
more intuitive to humans:

    middlewares := &MiddlewareStack
    middlewares.Use(middleware1)
    middlewares.Use(middleware2)
    handler := middlewares.Mount(mux)

In the example above, `middleware1` handles the request first, then
`middleware2`, then it's sent into the serve mux.

I also convert the River UI handler's `stripPrefix` function over to
middleware to demonstrate what usage looks like.