Prefer use of `rivershared` utilities instead of vendored versions
Now that we have a `rivershared` package containing many common
utilities and features, we can remove the variants added to River UI in
favor of those, with the net result being less code duplication between
projects.

I also renamed `dbutil` to `pgxutil` because we have a `dbutil` in
`rivershared` with similar features, except that it works on executors
instead of Pgx transactions, so it was a little confusing that they seem
to be the same package but do slightly variant things. We'll have to
keep `pgxutil` in River UI only so that we don't add a hard Pgx
dependency in the main River project (where a Pgx dependency should only
come in through the Pgx driver).