Upgrade to Go 1.23 + golangci-lint 1.60 (#112)
As we're doing in other River repos, upgrade to Go 1.23 and
golangci-lint 1.60 (for compatibility with Go 1.23).

Just in case we get this handler Go API working at some point, I've
added Go 1.23 as an additional version in the CI matrix, and left the
minimum Go version as 1.22 in `go.mod`. We can always change this later
in case it becomes painful to support both.