Add `dist` directory with `.gitkeep` placeholder file (#158)
I just noticed that on a fresh install, you can't run tests or Go lint
because `go:embed` fails because the repository has no `dist/`
directory.

This can be fixed obviously by building a JavaScript distribution, but
this isn't super obvious from the error messages. Here, add a `dist/` to
the repo by keeping the empty directory around with a `.gitkeep` file.