Change `HOST` to `RIVER_HOST`
I realized after merging #157 that it's probably not appropriate to use
the env var `HOST` because it's so ubiquitous and easily conflated with
a value intended for a different, non-River use. For example, `HOST` is
set automatically on every computer running Mac OS, and likely other
operating systems as well.

Since we're prefixing variables with `RIVER_` for other purposes like
`RIVER_LOG_LEVEL` in #183 anyway, change this to `RIVER_HOST`.

I think we should probably prefix some of the other env vars that
River UI uses as well (e.g. `DEV` should maybe be `RIVER_DEV`), but
after starting to do that I gave up on it for now since there were a
couple where the right thing to do (prefix versus not) wasn't totally
clear.