Simplify env file setup somewhat
Switch to `dotenv_if_exists` direnv command, and make it clear by
comment that this is a direnv built-in, so the dotenv executable isn't
actually needed.

Also, make `.env.example` just `.env.local` instead and commit it
straight to the repo. This has the advantage of needing fewer setup
instructions in `development.md`, but also means that if we ever need to
add any new default development vars, we can just put them in
`.envrc.local` and everybody will get them automatically.

[1] https://direnv.net/man/direnv-stdlib.1.html#codedotenvifexists-ltdotenvpathgtcode