Drop dotenv requirement + simplifying setup somewhat
Just a minor prospective improvement for getting the project
bootstrapped. The `.envrc` in here right now sources `.env.local` via
dotenv, which means that both direnv _and_ dotenv are technically
require to run it (and personally, I never use dotenv anymore).

Direnv has a handy built-in called `source_env_if_exists` that basically
does exactly the same thing, and we can use that instead of falling back
to dotenv. It's also a bit more succinct because the "if exist" is built
right in so no shell checks are necessary.

Also, make `.env.example` just `.envrc.local` instead and commit it
straight to the repo. This has the advantage of needing fewer setup
instructions in `development.md`, but also means that if we ever need to
add any new default development vars, we can just put them in
`.envrc.local` and everybody will get them automatically.