fix frontend type issues with split Job type
Issue #347 details how #344 broke some key parts of the frontend due to
missing fields in the job list payloads (errors, logs, and metadata were
removed from the job payload for the list API). The root issue is that
the TypeScript types were not adjusted to reflect the new split Job
type, so several parts of the code made reference to things they
could no longer access (and didn't actually need anyway). In particular
this was true for deserializing job payloads where the same logic was
applied to all job responses.

To fix this, update the types to have a separate `JobMinimal` and `Job`
type just as the API backend does. Update all code that uses them to
ensure the correct one is utilized. It's only list view stuff that needs
`JobMinimal` for now.

Fixes #347.