Add health check command and document health checks (#390)
# Motivation
I wanted to use RiverUI container as a standalone and not embed it.
I'm running it on AWS ECS Fargate which sits behind an AWS ALB

# Changes Summary
1. Allow the health endpoint to be requested even when Basic Auth is applied (ignore any path containing `/api/health-checks`)
2. Added a CLI flag to check the health 
3. Documented the existing health check endpoint and the new CLI flag

No. 2 is necessary since AWS ECS tasks detect the container health via commands like in docker compose. And the AWS ALB will need the health endpoint to determine if it should route requests to the target group. So both approaches are needed for proper health checking across the stack.

**Why make the CLI request the health endpoint?**
I want  to avoid pulling in another "container" dependency like `curl` or increase its size.
I thought of making the CLI apply the health check logic directly but that doesn't check the actual server health. It will check that the binary can connect to the DB. Which is useless.