Tweak a few test case names to use `TestCaseCamelCaseConvention`
Just a couple tweaks to test names to use `CamelCaseConvention` instead
of "words with spaces convention". The former is used universally in
other River projects and is kind of nice because (1) a name can be
yanked by double clicking it or a two-letter Vim motion like `yw`, and
(2) the test names are the same in the editor as they are when Go is
running. Using spaces, Go mutates test names to include underscores like
`words_with_spaces_convention`.