rename Server + ServerOpts to Handler + HandlerOpts (#400)
Although it must be "started" for its background caching routine to
function, this type ultimately still functions as an `http.Handler`
which can be embedded into an `http.ServeMux` or other
handler-compatible router. It does _not_ function as an `http.Server`
and the prior naming caused confusion in this regard.

Bite the bullet on the breaking rename while we already have other
breaking changes going out for the endpoints concept.